requirements.txt
langchain
langgraph
langchain_community
pydantic
uvicorn
fastapi
youtube-search-python
langchain_openai
langchain_together
python-telegram-bot==20.7
python-dotenv==1.0.0
PyYAML==6.0.1
asyncio
aiofiles==23.2.1
